+++ nuvla/job_engine/connector/k8s_driver.py
@@ -576,7 +576,8 @@ users:
          'Job',
          'CronJob',
          'StatefulSet',
-         'DaemonSet']
+         'DaemonSet',
+         'Pod']
 
     @should_connect
     def log(self, component: str, since: datetime, num_lines: int,